Save located virtual services in handler for status.
  Instead of looping through all virtual services again.
